Rewrote the code all in JS
