--- AlphaVote/Alphaglitch-refined/backend/.flaskenv
@@ -1,2 +1,0 @@
-FLASK_APP=app.py
-FLASK_ENV=development
