--- AlphaVote/Alphaglitch-refined/backend/Procfile
@@ -1,2 +1,0 @@
-web: gunicorn wsgi:application
-release: flask db upgrade
