--- AlphaVote/Alphaglitch-refined/backend/wsgi.py
@@ -1,3 +1,0 @@
-from app import app
-
-application = app
