--- AlphaVote/Alphaglitch-refined/frontend/src/Home.js
@@ -1,7 +1,0 @@
-import ElectionList from "./sections/ElectionList";
-
-const Home = () => {
-  return <ElectionList />;
-};
-
-export default Home;
