--- AlphaVote/Alphaglitch-refined/frontend/src/Mock.js
@@ -1,100 +1,0 @@
-const data = [
-  {
-    elect_img:
-      "https://i.postimg.cc/0jsrn6mv/istockphoto-1034363382-612x612.jpg",
-    elect_tit: "Best Cryptocurrency",
-    elect_desc: "vote which cryptocurrency is best.",
-    elect_cand: [
-      {
-        cand_img: "https://i.postimg.cc/PxGqfLSD/choice.png",
-        cand_det: "ChoiceCoin",
-        cand_score: 35,
-      },
-      {
-        cand_img: "https://i.postimg.cc/gjHjTYqZ/download-1.png",
-        cand_det: "Bitcoin",
-        cand_score: 18,
-      },
-      {
-        cand_img: "https://i.postimg.cc/rmWK5GDg/ethereum-eth-logo.png",
-        cand_det: "Ethereum",
-        cand_score: 12,
-      },
-      {
-        cand_img: "https://i.postimg.cc/fyqFcQL7/doge.png",
-        cand_det: "Dogecoin",
-        cand_score: 5,
-      },
-    ],
-  },
-  {
-    elect_img:
-      "https://i.postimg.cc/jdPgKb8p/baea7d2229c8c6fd735ad9ddabf44a0e-removebg-preview2.png",
-    elect_tit: "US Presidential Elections",
-    elect_desc:
-      "The election of the president and the vice president of the United States is an indirect election in which citizens of the United States who are registered to vote in one of the fifty U.S. states or in Washington, D.C., cast ballots for members of the Electoral College.",
-    elect_cand: [
-      {
-        cand_img: "https://i.postimg.cc/C5WLz3K1/1.jpg",
-        cand_det: "Joe Biden",
-        cand_score: 900,
-      },
-      {
-        cand_img: "https://i.postimg.cc/7Y2ZpZBT/2.jpg",
-        cand_det: "Donald Trump",
-        cand_score: 760,
-      },
-      {
-        cand_img: "https://i.postimg.cc/bN9kKgVn/3.jpg",
-        cand_det: "Jo Jorgensen",
-        cand_score: 230,
-      },
-      {
-        cand_img: "https://i.postimg.cc/8zkMDvx0/4.jpg",
-        cand_det: "Howie Hawkins",
-        cand_score: 650,
-      },
-    ],
-  },
-  {
-    elect_img: "https://i.postimg.cc/brMTP6bN/4-c6d96be5.jpg",
-    elect_tit: "California gubernatorial recall election",
-    elect_desc: "Special recall election for the state of California",
-    elect_cand: [
-      {
-        cand_img: "",
-        cand_det: "Yes",
-        cand_score: 70,
-      },
-      {
-        cand_img: "",
-        cand_det: "Nv",
-        cand_score: 21,
-      },
-      {
-        cand_img: "",
-        cand_det: "No",
-        cand_score: 101,
-      },
-    ],
-  },
-  {
-    elect_img: "https://i.postimg.cc/x85HbGZh/baking-1293986-340.webp",
-    elect_tit: "Legalization of Marijuana reform poll",
-    elect_desc: "Vote to legalize marijuana in the state of Arizona",
-    elect_cand: [
-      {
-        cand_img: "",
-        cand_det: "Yes it should be legalized",
-        cand_score: 232,
-      },
-      {
-        cand_img: "",
-        cand_det: "No it shouldn't.",
-        cand_score: 43,
-      },
-    ],
-  },
-];
-
-export default data;
