--- AlphaVote/Alphaglitch-refined/frontend/src/constants.js
@@ -1,5 +1,0 @@
-const URL = "https://choice-coin-api.herokuapp.com";
-// const URL = "http://localhost:5000";
-const ASSET_ID = 21364625;
-
-export { URL, ASSET_ID };
