--- AlphaVote/Alphaglitch-refined/frontend/src/elections/index.js
@@ -1,11 +1,0 @@
-import StartElection from "../sections/StartElection";
-
-const Elections = () => {
-  return (
-    <>
-      <StartElection />
-    </>
-  );
-};
-
-export default Elections;
