--- AlphaVote/Alphaglitch-refined/frontend/src/sections/StartElection.js
@@ -1,172 +1,0 @@
-import axios from "axios";
-import "../styles/startelect.css";
-import { URL } from "../constants";
-import { useQuery } from "react-query";
-import { NavLink } from "react-router-dom";
-import { BarLoader } from "react-spinners";
-import { useDispatch, useSelector } from "react-redux";
-
-const StartElection = () => {
-  const darkTheme = useSelector((state) => state.status.darkTheme);
-
-  // wallet-type & address
-  const walletAddress = localStorage.getItem("address");
-  const headers = { "X-Wallet-Address": walletAddress };
-  const { isLoading, error, data } = useQuery("elections", () =>
-    axios
-      .get(`${URL}/elections/mine`, { headers })
-      .then((response) => response.data.data)
-  );
-
-  const isWalletConnected =
-    localStorage.getItem("wallet-type") === null ? false : true;
-  const dispatch = useDispatch();
-
-  return (
-    <div className="stt_elt">
-      <div className="stt_elt_inn">
-        <div className="stt_hd">Recently Created Elections</div>
-
-        <ul className="on_elt">
-          {isLoading && !error ? (
-            <div
-              style={{
-                width: "100%",
-                height: "100px",
-                display: "flex",
-                flexDirection: "column",
-                alignItems: "center",
-                fontWeight: "500",
-                color: "var(--wht)",
-                justifyContent: "center",
-                textTransform: "uppercase",
-              }}
-            >
-              <p style={{ opacity: 0.5, marginBottom: "30px" }}>
-                Retrieving your Elections
-              </p>
-
-              <BarLoader
-                color={darkTheme ? "#eee" : "#555"}
-                loading={true}
-                size={10}
-                speedMultiplier={0.5}
-              />
-            </div>
-          ) : !!(data?.length > 0) ? (
-            data?.map((item, index) => {
-              return (
-                <li
-                  key={`${index}`}
-                  className="elt_item"
-                  onClick={() =>
-                    dispatch({ type: "popupElection", payload: item })
-                  }
-                  style={{
-                    borderBottom: index === 0 && "var(--list-border)",
-                  }}
-                >
-                  <div className="elt_img_cont">
-                    <img src={item.process_image} alt="" />
-                  </div>
-                  <div className="elt_det">
-                    <p className="e_det_main">{`${item.title}`}</p>
-                    <p className="e_det_sub">{`${
-                      !!item.description
-                        ? item.description
-                        : "No description provided"
-                    }`}</p>
-                  </div>
-                  <div className="pop_butt">
-                    <i className="uil uil-ellipsis-v"></i>
-                  </div>
-                </li>
-              );
-            })
-          ) : (
-            <>
-              <div
-                style={{
-                  flex: 1,
-                  opacity: 0.7,
-                  width: "100%",
-                  display: "flex",
-                  fontSize: "14px",
-                  color: "var(--wht)",
-                  letterSpacing: "1px",
-                  alignItems: "center",
-                  padding: "10px 20px",
-                  textAlign: "center",
-                  justifyContent: "center",
-                  textTransform: "uppercase",
-                }}
-              >
-                <p style={{ lineHeight: "20px" }}>
-                  Your Elections are displayed here once you connect your wallet
-                </p>
-              </div>
-            </>
-          )}
-        </ul>
-
-        <div
-          style={{
-            display: "flex",
-            width: "100%",
-            margin: "20px 0px 30px",
-            alignItems: "center",
-            flexDirection: "column",
-          }}
-        >
-          <i className="uil uil-arrow-break"></i>
-        </div>
-
-        <div className="stt_hd">do any of the following</div>
-        <div className="create_elect">
-          {
-            // If Wallet is connected this would be true
-            !!isWalletConnected ? (
-              <NavLink
-                style={{
-                  width: "100%",
-                  height: "100%",
-                  textAlign: "center",
-                  cursor: "pointer",
-                }}
-                to={`./create`}
-                key={"create"}
-              >
-                Create New Election
-              </NavLink>
-            ) : (
-              <div
-                style={{
-                  width: "100%",
-                  height: "100%",
-                  textAlign: "center",
-                  cursor: "pointer",
-                }}
-                onClick={() => {
-                  dispatch({ type: "modal_connect" });
-                }}
-              >
-                Create New Election
-              </div>
-            )
-          }
-        </div>
-        <div className="participate_elt">
-          <NavLink
-            style={{ width: "100%", height: "100%", textAlign: "center" }}
-            to={`../participate`}
-            key={"participate"}
-          >
-            Participate In Election
-          </NavLink>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default StartElection;
