+++ AlphaVote/Alphaglitch-refined/package.json
@@ -22,7 +22,6 @@
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
-    "react-spinners": "^0.11.0",
     "react-switch": "^6.0.0",
     "three": "^0.135.0",
     "web-vitals": "^1.0.1"
