+++ AlphaVote/Alphaglitch-refined/public/choice.png
