+++ AlphaVote/Alphaglitch-refined/public/manifest.json
