+++ AlphaVote/Alphaglitch-refined/src/MainPage.js
@@ -1,11 +1,7 @@
-import { NavLink, Route, Routes } from "react-router-dom";
-
-import Home from "./Home";
-import Transfer from "./transfer";
-import Elections from "./elections";
 import { useSelector } from "react-redux";
-import CreateElection from "./elections/create";
+import ElectionList from "./ElectionList";
 import { useWindowSize } from "@react-hook/window-size";
+import { NavLink, Route, Routes } from "react-router-dom";
 import TopNavigationBar from "./statics/TopNavigationBar";
 import BottomNavigationBar from "./statics/BottomNavigationBar";
 
@@ -23,11 +19,8 @@ const MainPage = () => {
       <TopNavigationBar darkTheme={darkTheme} />
 
       <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/home" element={<Home />} />
-        <Route path="/elections" element={<Elections />} />
-        <Route path="/transfer" element={<Transfer />} />
-        <Route path="/elections/create" element={<CreateElection />} />
+        <Route path="/" element={<ElectionList />} />
+        <Route path="/home" element={<ElectionList />} />
       </Routes>
 
       {width <= 850 && (
