+++ AlphaVote/Alphaglitch-refined/src/constants.js
@@ -1,0 +1,2 @@
+const ASSET_ID = 21364625;
+export { ASSET_ID };
