+++ AlphaVote/Alphaglitch-refined/src/statics/ScrollText.js
