+++ AlphaVote/Alphaglitch-refined/src/styles/reset.css
